ZooMe application
ZooMe is an application I developed in Android Studio for Android users. The application uses the user's camera to identify animals and displays information about them. The image recognition is based on a CNN algorithm that was learned on 5000 images and divided into five classes. I downloaded the images from Google's API. The information displayed to the user is pulled from Fire Base.